Add Search Higher Order Component
- This makes the individual search pages (Match/Range/UBRN) redundant, as those pages use almost identical functionality, we can move all the common logic (all of it) into a higher order component, which returns a component based upon the input settings
- Modify UBRN redux action to handle commonality between match + range
