+++ src/components/UBRNForm.js
@@ -7,7 +7,7 @@ class UBRNForm extends React.Component {
   render() {
     return (
       <form>
-        <TextInput ref={ip => (this.childTextInput = ip)} value={this.props.value} label="UBRN" id="UBRN" autoFocus onChange={this.props.onChange} /><br />
+        <TextInput ref={ip => (this.childTextInput = ip)} value={this.props.initialValues.id} label="UBRN" id="id" autoFocus onChange={this.props.onChange} /><br />
         <Button id="loginButton" size="wide" text="Search" onClick={!this.props.currentlySending ? this.props.onSubmit : null} ariaLabel="Login Button" type="submit" loading={this.props.currentlySending} />
         &nbsp;
         <Button id="clearButton" size="wide" text="Clear" onClick={this.props.onClear} ariaLabel="Clear Button" type="reset" />
@@ -22,7 +22,7 @@ UBRNForm.propTypes = {
   onSubmit: PropTypes.func.isRequired,
   onChange: PropTypes.func.isRequired,
   onClear: PropTypes.func.isRequired,
-  value: PropTypes.string.isRequired,
+  initialValues: PropTypes.object.isRequired,
 };
 
 export default UBRNForm;
